Extend algorithm, use getpwnam_r().
The nss_identity group functions now follow this logic:

A. If request is for a group by GID:
   1. If GID corresponds to a valid UID, obtain the passwd entry
      for this UID and construct a grp entry based upon that UID(GID) and
      pw_name.  Put the identified user into the group.
   2. Otherwise, construct a grp entry using the requested GID and and
      the group name "group_####", where "####" is the decimal ASCII
      encoding of the GID.

B. If request is for a group by name:
   1. If group name corresponds to a valid user name, obtain the passwd entry
      for this user and construct a grp entry based upon that UID(GID) and
      the requested group name.

   2. Otherwise, if group name matches "group_####", extract the #### as
      a GID and perform GID lookup (see ruleset A) using that GID.

   3. Otherwise, fail to discover the group.

Note that the group entries thus fabricated are not necessarily
reversible!  For example, if user "abc" has uid 1234, then a lookup on
gid=1234 will yield the group entry "abc::1234:abc".  A lookup on group
name="abc" will yield the same -- this is reversible.  However, a lookup
on "group_1234" will yield the group entry "group_1234:xx:1234:abc".
That cannot be reversed back to group_1234 because that name is
fabricated in the presence of a user "abc" with uid 1234.